api/pyproject.toml
[tool.poetry]
name = "cookiebreaks"
version = "1.0.0"
description = "Suite for managing the cookie break"
authors = ["George Kaye <georgejkaye@gmail.com>"]
license = "GPL-v3.0-or-later"
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11"
Jinja2 = "^3.1.2"
PyYAML = "^6.0"
psycopg-binary = "^3.1.9"
arrow = "^1.2.3"
icalendar = "^5.0.7"
fastapi = "^0.99.1"
uvicorn = "^0.22.0"
psycopg2 = "^2.9.6"
python-dotenv = "^1.0.0"
python-multipart = "^0.0.6"
python-jose = { extras = ["cryptography"], version = "^3.3.0" }
passlib = "^1.7.4"


[tool.poetry.group.dev.dependencies]
mypy = "^1.4.1"
black = "^23.3.0"
flake8 = "^6.0.0"
types-python-jose = "^3.3.4.8"
types-passlib = "^1.7.7.12"
types-PyYAML = "^6.0.12.10"

[tool.poetry.scripts]
start = "cookiebreaks.api.main:start"
dev = "cookiebreaks.api.main:dev"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
